# Back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Seguros TM",
        "description": "API para gestão de seguros, clientes, apólices e sinistros",
        "contact": {
            "email": "suporte@segurostm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/cliente/sinistros": {
            "get": {
                "tags": [
                    "Cliente - Sinistros"
                ],
                "summary": "Listar sinistros do cliente",
                "description": "Retorna todos os sinistros registrados pelo cliente autenticado",
                "operationId": "83beb29a87822b0fb424d3164442cbcd",
                "responses": {
                    "200": {
                        "description": "Lista de sinistros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_sinistro": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "numero_sinistro": {
                                                        "type": "string",
                                                        "example": "SIN-2024-001"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "em_analise"
                                                    },
                                                    "data_comunicacao": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "descricao_ocorrido": {
                                                        "type": "string"
                                                    },
                                                    "valor_estimado_dano": {
                                                        "type": "number"
                                                    },
                                                    "apolice": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cliente - Sinistros"
                ],
                "summary": "Registrar novo sinistro",
                "description": "Registra uma ocorrência de sinistro para uma apólice ativa",
                "operationId": "1331a96a2a8d6b9ca71c91bafcadebb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "apolice_id",
                                    "data_ocorrencia",
                                    "descricao_ocorrido",
                                    "valor_estimado_dano"
                                ],
                                "properties": {
                                    "apolice_id": {
                                        "description": "ID da apólice",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_ocorrencia": {
                                        "description": "Data em que ocorreu o sinistro",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "hora_ocorrencia": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30:00"
                                    },
                                    "local_ocorrencia": {
                                        "type": "string",
                                        "example": "Av. Julius Nyerere, Maputo"
                                    },
                                    "descricao_ocorrido": {
                                        "type": "string",
                                        "example": "Colisão traseira em semáforo"
                                    },
                                    "valor_estimado_dano": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "houve_vitimas": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "boletim_ocorrencia": {
                                        "type": "string",
                                        "example": "BO-2024-12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sinistro registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sinistro registado com sucesso"
                                        },
                                        "sinistro": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao registrar sinistro"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT. Use o token retornado no endpoint de login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}